Some success in automatically testing the codegen in Visual Studio 2013.
The basic premise seems sound. We should be able to declare a series of expressions to test, compile them, and analyze the opcodes in the resulting assembly code. Having a program evaluate its own assembly is a little... unconventional, but should allow us to declare our success criterion inline with the code to be tested.