BREAKING CHANGE: Greatly reduce scope of matrix overloads to only those operations which have a well defined algebraic meaning.
This changes the meaning of the expression `a*b` for matrices a, b from the original element-wise product to the more widely used matrix product, while at the same time deprecating the variadic function `mul(...)` which previously provided the matrix product.

The macro `LINALG_LEGACY_MATRIX_PRODUCT` can be defined to retain `mul(...)` instead of the `operator *` overloads to facilitate codebases transitioning from previous versions, but either way the element-wise product is removed. It can be manually invoked using `zip`.